+++ docs/reference/workspace-layout.md
@@ -118,8 +118,8 @@ Key paths:
   `overlay/session_state/frozen_annotation.rs`, frozen capture/export lifecycle state lives under
   `overlay/session_state/frozen_capture.rs`, and scroll-capture session state lives under
   `overlay/session_state/scroll_capture.rs`
-- `packages/rsnap-overlay/src/frozen_edit.rs`: Rust-owned frozen overlay edit session state machine;
-  edit geometry, text hit bounds, and movement clamping live under `frozen_edit/geometry.rs`
+- `packages/rsnap-overlay/src/frozen_edit.rs`: Rust-owned frozen overlay edit session state machine,
+  including its private edit geometry, text hit bounds, and movement clamping policy
 - `packages/rsnap-overlay/src/frozen_export.rs`: Rust-owned frozen-overlay export compositor;
   export element schema lives under `frozen_export/model.rs`, and stroke rasterization lives under
   `frozen_export/stroke_raster.rs`
