+++ packages/rsnap-overlay/src/overlay/worker_runtime/live_sampling.rs
@@ -10,6 +10,60 @@ use crate::overlay::{
 use crate::state::{LoupeSample, WindowHit};
 
 impl OverlaySession {
+	pub(in crate::overlay) fn maybe_timeout_pending_click_hit_test(
+		&mut self,
+		now: Instant,
+	) -> bool {
+		let Some(request_id) = self.pending_click_hit_test_request_id else {
+			self.pending_click_hit_test_requested_at = None;
+
+			return false;
+		};
+		let Some(requested_at) = self.pending_click_hit_test_requested_at else {
+			return false;
+		};
+		let Some(elapsed) = now.checked_duration_since(requested_at) else {
+			return false;
+		};
+
+		if elapsed < PENDING_CLICK_HIT_TEST_TIMEOUT {
+			return false;
+		}
+
+		self.pending_click_hit_test_request_id = None;
+		self.pending_click_hit_test_requested_at = None;
+
+		tracing::warn!(
+			request_id,
+			elapsed_ms = elapsed.as_millis(),
+			timeout_ms = PENDING_CLICK_HIT_TEST_TIMEOUT.as_millis(),
+			"Pending click hit test timed out."
+		);
+
+		if !matches!(self.state.mode, OverlayMode::Live) {
+			return true;
+		}
+
+		let LiveCaptureInteraction::PressPending { monitor, .. } = self.live_capture_interaction
+		else {
+			return true;
+		};
+		let next_interaction = self
+			.state
+			.cursor
+			.filter(|cursor| monitor.contains(*cursor))
+			.and_then(|cursor| {
+				self.live_capture_target_from_snapshot(monitor, cursor)
+					.map(|target| LiveCaptureInteraction::HoverWindow { monitor, target })
+			})
+			.unwrap_or(LiveCaptureInteraction::Idle);
+
+		self.set_live_capture_interaction(next_interaction);
+		self.request_redraw_for_monitor(monitor);
+
+		true
+	}
+
 	pub(in crate::overlay) fn request_live_samples_for_cursor(
 		&mut self,
 		monitor: MonitorRect,
