+++ packages/rsnap-overlay/src/scroll_capture.rs
@@ -2,7 +2,6 @@ pub mod bench_support;
 
 mod downward_candidates;
 mod downward_committed_recovery;
-mod downward_followup;
 mod downward_recovery;
 mod downward_registration;
 mod downward_resolution;
